+++ pycosio/storage/s3.py
@@ -67,13 +67,14 @@ class _S3System(_SystemBase):
         self._session = None
         _SystemBase.__init__(self, *args, **kwargs)
 
-    def copy(self, src, dst):
+    def copy(self, src, dst, other_system=None):
         """
         Copy object of the same storage.
 
         Args:
             src (str): Path or URL.
             dst (str): Path or URL.
+            other_system (pycosio._core.io_system.SystemBase subclass): Unused.
         """
         copy_source = self.get_client_kwargs(src)
         copy_destination = self.get_client_kwargs(dst)